Become a provider (#128)
IMPORTANT NOTE: we are planning to use a toast (or spinner and text) and
activity log to show transaction status. Therefore, I have left the old
tx status in the UI (with broken styling) so we can test. We will remove
this once the new toast/activity log is in place.


![image](https://github.com/LibertyDSNP/provider-dashboard/assets/1135694/0daab5c2-ece3-45cd-91b6-3814782f1462)