update file structure (#233)
## Goal

The goal of this PR is to update the file structure to something more
readable

Closes #211 

## Discussion

- created atoms, features, and pages sub directories
- grouped features component into feature specific sub directories

- an **atom** is something like a button, card or modal - very general
- a **feature** is a component or group of components that make up an
app-specific feature. Rule of thumb to be included in a feature
directory is if you are a child component that is only used in that
specific feature. If a child component is used in more than one page or
feature and not an atom, they become their own feature.
- **pages** are used for components that define a whole page's content
but are used conditionally on the same path.

## Checklist

- [ ] PR Self-Review and Commenting
- [ ] Docs updated
- [ ] Tests added

## How To Test the Changes

1. Clone the pr branch
2. `npm run dev`
3. check that the app still runs as expected
4. navigate through the file structure